+++ service/history/events/notifier.go
@@ -41,8 +41,8 @@ import (
 	"go.temporal.io/server/common/definition"
 	"go.temporal.io/server/common/metrics"
 	"go.temporal.io/server/common/namespace"
+	"go.temporal.io/server/common/persistence/transitionhistory"
 	"go.temporal.io/server/common/persistence/versionhistory"
-	"go.temporal.io/server/service/history/workflow"
 )
 
 const (
@@ -119,7 +119,7 @@ func NewNotification(
 		WorkflowState:          workflowState,
 		WorkflowStatus:         workflowStatus,
 		VersionHistories:       versionhistory.CopyVersionHistories(versionHistories),
-		TransitionHistory:      workflow.CopyVersionedTransitions(transitionHistory),
+		TransitionHistory:      transitionhistory.CopyVersionedTransitions(transitionHistory),
 	}
 }
