+++ service/history/workflow/state_transition_history.go
@@ -143,25 +143,3 @@ func CompareVersionedTransition(
 
 	return 0
 }
-
-func CopyVersionedTransitions(
-	transitions []*persistencespb.VersionedTransition,
-) []*persistencespb.VersionedTransition {
-	if transitions == nil {
-		return nil
-	}
-	copied := make([]*persistencespb.VersionedTransition, len(transitions))
-	for i, t := range transitions {
-		copied[i] = CopyVersionedTransition(t)
-	}
-	return copied
-}
-
-func CopyVersionedTransition(
-	transition *persistencespb.VersionedTransition,
-) *persistencespb.VersionedTransition {
-	if transition == nil {
-		return nil
-	}
-	return common.CloneProto(transition)
-}
